chore(lint): clear remaining ruff backlog + config intentional ignores
Mechanical fixes from `ruff --fix --unsafe-fixes` (RUF010 redundant
`int(round())`, RUF015 unused tuple elements, RUF007 zip → pairwise,
B905 zip strict=, I001 import order).

Wrapped 4 long-line tests + 1 long-line src line by hand (E501).

Configured ruff to ignore three classes of finding that catch nothing
real in this project:
- RUF001/002/003: en-dashes for ranges, `×` for multiplication, `∨` for
  logical-or in math comments, and the NO-BREAK SPACE in Amazon's
  "Sold by:" strip-set are all intentional typography. ASCII-fying them
  degrades the docs.
- Per-file E501 on Alembic-generated migrations + SQLAlchemy field
  declarations + raw-SQL view defs.
- Per-file E702 on tests: the compact
  `session.add(x); session.commit(); session.refresh(x)` fixture idiom
  is used consistently across the suite.

`ruff check` now reports zero findings; pytest still 285 passed.

+++ src/book_alerter/stats.py
@@ -175,7 +175,7 @@ def _percentile_rank(sorted_totals: list[int], value: int) -> int:
     if not sorted_totals:
         return 0
     below = bisect_right(sorted_totals, value)
-    return int(round((below / len(sorted_totals)) * 100))
+    return round((below / len(sorted_totals)) * 100)
 
 
 def _imputed_shipping(
